added data size in table

+++ bin/storage/storage.js
@@ -7,35 +7,30 @@
 
 import { promises as fsPromises } from "fs";
 import path from "path";
-import chalk from "chalk";
 
 export class Storage {
-  clearDirectoryAndGetSize = async (fullPath) => {
+  clearDirectoryAndGetSize = async (fullPath, deletion) => {
     try {
       const stats = await fsPromises.stat(fullPath);
       if (stats.isDirectory()) {
-        const size = await this.calculateDirectorySize(fullPath);
+        const size = await this.calculateDirectorySize(fullPath, deletion);
         return size;
       } else {
         try {
-          await fsPromises.unlink(fullPath);
+          if (deletion) {
+            await fsPromises.unlink(fullPath);
+          }
           return stats.size;
         } catch (err) {
-          console.log(`Skipping Busy files ${fullPath}`, err.message);
           return 0;
         }
       }
     } catch (error) {
-      console.error(
-        chalk.redBright(
-          `Error retrieving storage information for ${fullPath}: ${error}`
-        )
-      );
       return 0;
     }
   };
 
-  calculateDirectorySize = async (dirPath) => {
+  calculateDirectorySize = async (dirPath, deletion) => {
     let totalSize = 0;
     const files = await fsPromises.readdir(dirPath, { withHidden: true });
 
@@ -43,14 +38,14 @@ export class Storage {
       files.flatMap(async (file) => {
         const fullPath = path.join(dirPath, file);
 
-        return await this.clearDirectoryAndGetSize(fullPath);
+        return await this.clearDirectoryAndGetSize(fullPath, deletion);
       })
     );
     sizeList?.forEach((size) => (totalSize += size));
-    try {
-      await fsPromises.rmdir(dirPath);
-    } catch (err) {
-      console.error(`Error deleting directory ${dirPath}`, err.message);
+    if (deletion) {
+      try {
+        await fsPromises.rmdir(dirPath);
+      } catch (err) {}
     }
     return totalSize;
   };
